Fix beam search: f64 precision, tie-break, and Viterbi backtrack
The beam kernel disagreed with the notebook's numba reference by up to
2.0 ft max / 0.34 ft mean on real wells, despite the existing test suite
passing — the CubeCL CPU-interpreted tests only ever compared the kernel
against this crate's own (equally wrong) reference, never against numba.

Three independent bugs, found by tracing the real, currently-active
_beam_jit/_smooth (cell 37 shadows cell 12's plain-Python versions, and
run_beam_ensemble resolves the shadowed ones at call time):

- The kernel ran in f32 while the reference runs in f64; over a
  thousands-of-steps dynamic program, rounding differences at near-tie
  costs flip which candidate wins. Beam search now runs in f64
  end-to-end (kernel, host buffers, Python bindings).
- The GR smoother implemented Savitzky-Golay quadratic fitting, but
  _smooth is a plain centred rolling mean with min_periods=1 semantics
  (13 of the notebook's 14 configs smooth, so this alone accounted for
  most of the gap).
- The ranking step's tie-break used type-well index; _beam_jit ties by
  candidate generation order (a hand-rolled selection sort over an array
  built in (beam slot, move) order), so it needed the same tie-break.

The largest fix: _beam_jit is a Viterbi backtrack (every step keeps
per-slot parent pointers, then walks backward from the cheapest FINAL
path once the whole sequence has run), not greedy per-step emission.
A candidate that isn't cheapest at step t can still win the whole
search later, so emitting the step-wise argmin is a different
algorithm. The kernel now scatters (index, parent) history into a
global-memory buffer per step and backtracks from one unit per cube
after the sequence completes.

Result: max/mean abs diff vs. numba dropped from 2.0/0.34 ft to
0.002/0.0015 ft (float64 summation-order noise) across 6 real wells,
and 0/0 on a real HIP run of the exact case the CPU-interpreted test
uses. That CPU-interpreted test is now ignored with an explanation:
the interpreter doesn't correctly reconverge units around the kernel's
divergent single-unit backtrack, confirmed by the same scenario passing
bit-exact on real HIP hardware (new beam_backtrack_matches_reference_on_hip
test in tests/rocm_parity.rs).

Bumped to 0.2.1 (Cargo.toml was still 0.1.0, out of sync with
pyproject.toml's 0.2.0; both now match the release version).

+++ python/rog2_pf/__init__.py
@@ -258,9 +258,12 @@ def prepare_beam_well(hw, tw):
     )
 
     well = {
-        "gr": gr_all[ev.index].astype(np.float32),
-        "tw_tvt": tw_tvt.astype(np.float32),
-        "tw_gr": tw_gr.astype(np.float32),
+        # Kept at f64: the beam search's dynamic program makes strict-inequality
+        # cost comparisons over thousands of sequential steps, and the notebook's
+        # numpy reference runs entirely in f64 (see beam_host's launch macros).
+        "gr": gr_all[ev.index],
+        "tw_tvt": tw_tvt,
+        "tw_gr": tw_gr,
         "last_tvt": float(kn.iloc[-1].TVT_input),
     }
     return well, ev.index.values
@@ -323,15 +326,15 @@ def beam_search(hgr, tw_tvt, tw_gr, last_tvt, bs=10, mc=20.0, es=144.0, r=2, **k
     ``tw_tvt`` would silently give a wrong answer; the notebook's callers all
     pass ``tw.sort_values('TVT')`` anyway, so this is a no-op for them.
     """
-    hgr = np.asarray(hgr, dtype=np.float32)
+    hgr = np.asarray(hgr, dtype=np.float64)
     if len(hgr) == 0:
         return np.array([float(last_tvt)])
 
     order = np.argsort(np.asarray(tw_tvt, dtype=np.float64), kind="stable")
     well = {
         "gr": hgr,
-        "tw_tvt": np.asarray(tw_tvt, dtype=np.float32)[order],
-        "tw_gr": np.asarray(tw_gr, dtype=np.float32)[order],
+        "tw_tvt": np.asarray(tw_tvt, dtype=np.float64)[order],
+        "tw_gr": np.asarray(tw_gr, dtype=np.float64)[order],
         "last_tvt": float(last_tvt),
     }
     res = run_beam_batch(
